refactor: removed extra spaces from files
