+++ src/components/UI/Blog.module.css
@@ -8,7 +8,6 @@
     color: white;
     border: 1px solid black;
     background-color: rgba(63, 61, 59, 0.748);
-    ;
     color: black;
     z-index: 1;
 }
