+++ src/pages/BlogDetail.module.css
@@ -9,7 +9,6 @@
 }
 
 .detail p {
-    overflow: scroll;
     text-align: justify;
     display: block;
     height: max-content;
